refactor: implement GameDetailsPanel component and streamline GameView structure fix: logo showing up in the correct location on load.
